Conflicts:
	Sources/Internal/Render/Highlevel/Landscape.cpp
	Sources/Internal/Render/Material/NMaterialNames.cpp
	Sources/Internal/Render/Material/NMaterialNames.h
	Tools/ResourceEditor/Data/Materials/Shaders/Landscape/tilemask.fsh
	Tools/ResourceEditor/Data/Materials/Shaders/Landscape/tilemask.vsh

Former-commit-id: 44c141dfebefa6a322f53b491815456125664c0f

+++ Sources/Internal/Base/HashMap.h
@@ -154,7 +154,7 @@ HashMap<K, V>::HashMap(size_t _hashSize, V _defaultV)
 	, defaultV(_defaultV)
 {
 	// not 0 and power of 2
-	DVASSERT(IsPowerOf2(szTable));
+	DVASSERT(IsPowerOf2((int32)szTable));
 
 	table = new HashMapItem*[szTable];
 	for(size_t i = 0; i < szTable; ++i)
