+++ Tools/UIEditor/Classes/UI/PropertyGridWidgets/Custom/guidepropertygridwidget.h
@@ -53,7 +53,7 @@ protected:
     
 protected slots:
     void OnGuideMoved(GuideData* guideData);
-    void OnGuidePositionChanged(int value);
+    void OnGuidePositionChanged(double value);
 
 private:
     Ui::GuidePropertyGridWidget *ui;
